requirements.txt
streamlit
opencv-python
tensorflow
Pillow
st-pages
torch
torchvision>=0.16.2
detectron2>=0.6,<0.8